+++ internal/args/parse_test.go
@@ -3,6 +3,8 @@ package args
 import (
 	"reflect"
 	"testing"
+
+	"github.com/alecthomas/kong"
 )
 
 func TestParseOk(t *testing.T) {
@@ -63,11 +65,11 @@ func TestParseOk(t *testing.T) {
 		"tolerable filter": {
 			input: []string{
 				"exodus-rsync",
-				"--filter",
-				"+ */",
+				"--filter", "+ **/hi/**",
+				"--filter=-/_*",
 				"x",
 				"y"},
-			want: Config{Src: "x", Dest: "y", Filter: "+ */"}},
+			want: Config{Src: "x", Dest: "y", Filter: []string{"+ **/hi/**", "-/_*"}}},
 	}
 	for name, tc := range tests {
 		t.Run(name, func(t *testing.T) {
@@ -122,5 +124,17 @@ func TestDestPath(t *testing.T) {
 				t.Errorf("Config.DestPath() = %v, want %v", got, tt.want)
 			}
 		})
+	}
+}
+
+func TestStringMapDecodeError(t *testing.T) {
+	err := argStringMapper{}.Decode(
+		&kong.DecodeContext{Value: &kong.Value{}, Scan: &kong.Scanner{}},
+		reflect.Value{},
+	)
+
+	// Scan holds no tokens; no flag or flag value to decode.
+	if err.Error() != "flag : missing value" {
+		t.Fatalf("didn't get expected error, got %s", err.Error())
 	}
 }