+++ internal/walk/walk.go
@@ -112,13 +112,12 @@ func fillItems(ctx context.Context, in <-chan walkItem, c chan<- syncItemPrivate
 	}
 }
 
-func getSyncItems(ctx context.Context, path string, exclude []string,
-	onlyTheseFiles []string) <-chan syncItemPrivate {
+func getSyncItems(ctx context.Context, path string, exclude []string, include []string, onlyThese []string) <-chan syncItemPrivate {
 	c := make(chan syncItemPrivate, 10)
 	walkItemCh := make(chan walkItem, 10)
 
 	go func() {
-		err := walkDirWithLinks(ctx, path, exclude, onlyTheseFiles,
+		err := walkDirWithLinks(ctx, path, exclude, include, onlyThese,
 			func(path string, d fs.DirEntry, err error) error {
 				if err != nil {
 					return err
@@ -148,11 +147,10 @@ func getSyncItems(ctx context.Context, path string, exclude []string,
 
 // Walk will walk the directory tree at the given path and invoke a handler
 // for every discovered item eligible for sync.
-func Walk(ctx context.Context, path string, exclude []string,
-	onlyTheseFiles []string, handler SyncItemHandler) error {
+func Walk(ctx context.Context, path string, exclude []string, include []string, onlyThese []string, handler SyncItemHandler) error {
 	logger := log.FromContext(ctx)
 
-	for item := range getSyncItems(ctx, path, exclude, onlyTheseFiles) {
+	for item := range getSyncItems(ctx, path, exclude, include, onlyThese) {
 		logger.F("item", item).Debug("got item")
 
 		if ctx.Err() != nil {